+++ packages/floop_web/example/example.dart
@@ -1,71 +1,64 @@
-import 'package:flutter_web/material.dart';
-import 'package:floop_web/floop_web.dart';
-
-/// This example exists solely to satisfy dart publishing requirements.
-/// The real examples folder lies at the root of the project on Github.
-/// https://github.com/icatalud/floop
+import 'package:flutter/material.dart';
+import 'package:floop/floop.dart';
+import 'package:http/http.dart' as http;
 
 void main() {
-  floop['clicks'] = 0;
-  runApp(MaterialApp(
-      title: 'Clicker',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: Clicker()));
+  fetchImage();
+  runApp(MaterialApp(title: 'Fetch image', home: ImageDisplay()));
 }
 
-class Clicker extends StatelessWidget with Floop {
+var _fetching = false;
+
+fetchImage([String url = 'https://picsum.photos/300/200']) async {
+  if (_fetching) {
+    return;
+  }
+  _fetching = true;
+  floop['image'] = null; // Set to null while awaiting the response
+  final response = await http.get(url);
+
+  // The image is stored only when this is the last call to fetchImage,
+  floop['image'] = TransitionImage(Image.memory(response.bodyBytes));
+  _fetching = false;
+}
+
+// `extends FloopWidget` is equivalent to `...StatelessWidget with Floop`.
+class TransitionImage extends FloopWidget {
+  final Image image;
+  const TransitionImage(this.image);
+
   @override
   Widget buildWithFloop(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: Text(
-          floop['clicks'].toString(),
-          style: TextStyle(
-            color: Colors.red,
-            fontSize: 100,
-          ),
-        ),
-      ),
-      floatingActionButton: FloatingActionButton(
-          child: Icon(Icons.add), onPressed: () => floop['clicks']++),
-    );
+    // Opacity transitions from 0 to 1 in 1.5 seconds.
+    return Opacity(opacity: transition(1500), child: image);
   }
 }
 
-// The following are alternative implementations.
-
-class ClickerStateful extends StatefulWidget {
-  @override
-  State<StatefulWidget> createState() => ClickerState();
-}
-
-class ClickerState extends State<ClickerStateful> with FloopStateMixin {
+class ImageDisplay extends StatelessWidget with Floop {
   @override
   Widget buildWithFloop(BuildContext context) {
     return Scaffold(
-      body: Center(
-          child: Text(floop['clicks'].toString(),
-              style: TextStyle(
-                color: Colors.red,
-                fontSize: 100,
-              ))),
+      // `floop['image']` is null while fetching an image. When the
+      // imaged is downloaded, an image widget is stored on `floop['image']`
+      // and the widget automatically updates.
+      body: floop['image'] == null
+          ? Center(
+              child: Text(
+                'Loading...',
+                textScaleFactor: 2,
+              ),
+            )
+          : Align(
+              alignment: Alignment(0, transition(2000, delayMillis: 800) - 1),
+              child: floop['image']),
       floatingActionButton: FloatingActionButton(
-          child: Icon(Icons.add), onPressed: () => floop['clicks']++),
-    );
-  }
-}
-
-// Simplest example.
-
-class SimpleClicker extends StatelessWidget with Floop {
-  @override
-  Widget buildWithFloop(BuildContext context) {
-    return Scaffold(
-      body: Center(child: Text(floop['clicks'].toString())),
-      floatingActionButton: FloatingActionButton(
-          child: Icon(Icons.add), onPressed: () => floop['clicks']++),
+        child: Icon(Icons.refresh),
+        onPressed: () async {
+          await fetchImage();
+          // Restart context transitions after image has been loaded.
+          Transitions.restart(context: context);
+        },
+      ),
     );
   }
 }