+++ packages/floop_web/lib/src/flutter_import.dart
@@ -1,12 +1,16 @@
-export 'package:flutter_web/widgets.dart'
-    show
-        Element,
-        BuildContext,
-        Widget,
-        StatelessWidget,
-        StatelessElement,
-        StatefulWidget,
-        State,
-        // annotations
-        mustCallSuper,
-        visibleForTesting;
+export 'package:flutter_web/widgets.dart';
+// show
+//     Element,
+//     Key,
+//     LocalKey,
+//     UniqueKey,
+//     BuildContext,
+//     Widget,
+//     StatelessWidget,
+//     StatelessElement,
+//     StatefulWidget,
+//     StatefulElement,
+//     State,
+//     // annotations
+//     mustCallSuper,
+//     visibleForTesting;